refactor: modularize cmake structure
