+++ InvokeInvoiceSystem.includes
@@ -1,2 +1,3 @@
 include
+libs/passlib/include
 out/build/x64-debug/vcpkg_installed/x64-windows/include
