+++ include/pch.h
@@ -23,7 +23,7 @@
 #include <bsoncxx/builder/basic/kvp.hpp>  
 #include "System/Database/MongoDBDataManager.h"
 // Include bcrypt library  
-#include "Accounts/PasswordHashing/bcrypt.h"  
+#include "passlib/bcrypt.h"
 
 //include pdf table
 extern "C" {
