+++ libs/passlib/src/openbsd.h
